Add files via upload
Build a circuit using your Raspberry Pi that causes an LED to blink when a push button is NOT pressed. However, the LED should stay on continually when the push button IS pressed.

Your video should show the LED blinking when the push button is not pressed, and it should show that the LED is constantly on while the button is pressed.